separaioned helpers and utils
