+++ .gitignore
@@ -5,3 +5,5 @@
 
 node_modules/
 public/bundle.js
+
+.DS_Store
