+++ src/utils/chat.js
@@ -1,0 +1,16 @@
+// Stole this function to send chat messages
+let lastMessage = 0;
+export default function chat(chatMessage) {
+  const limit = 500 + 10;
+  const now = new Date();
+  const timeDiff = now - lastMessage;
+  if (timeDiff > limit) {
+    tagpro.socket.emit('chat', {
+      message: chatMessage,
+      toAll: 0,
+    });
+    lastMessage = new Date();
+  } else if (timeDiff >= 0) {
+    setTimeout(chat, limit - timeDiff, chatMessage);
+  }
+}
