+++ src/utils/movement.js
@@ -1,0 +1,26 @@
+// Sends key events to move to a destination.
+export default function move(destination) {
+  // TODO: address deadband variable with a comment
+  const deadband = 4;
+  if (destination.x > deadband) {
+    tagpro.sendKeyPress('left', true);
+    tagpro.sendKeyPress('right', false);
+  } else if (destination.x < -deadband) {
+    tagpro.sendKeyPress('right', true);
+    tagpro.sendKeyPress('left', false);
+  } else {
+    tagpro.sendKeyPress('right', true);
+    tagpro.sendKeyPress('left', true);
+  }
+
+  if (destination.y > deadband) {
+    tagpro.sendKeyPress('up', true);
+    tagpro.sendKeyPress('down', false);
+  } else if (destination.y < -deadband) {
+    tagpro.sendKeyPress('down', true);
+    tagpro.sendKeyPress('up', false);
+  } else {
+    tagpro.sendKeyPress('up', true);
+    tagpro.sendKeyPress('down', true);
+  }
+}
